created the 'QuizResults' component with initial styles but without any actual logic yet
