+++ src/components/UI/quiz_results/QuizResults.module.css
@@ -1,0 +1,10 @@
+.quiz_results_block {
+    margin: 10px 0 10px 10px;
+    font-size: 18px
+}
+.quiz_results_block p, .quiz_results_block h3 {
+    margin-bottom: 10px
+}
+.quiz_results_block p:last-of-type {
+    margin-bottom: 50px
+}
